32-bit interface to hw accel

+++ services/interfaces/jd_accel.h
@@ -8,6 +8,6 @@ void acc_hw_init(void);
 
 void acc_hw_sleep(void);
 
-void acc_hw_get(int16_t sample[3]);
+void acc_hw_get(int32_t sample[3]);
 
 #endif
